backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
pymongo==4.7.2
beanie==1.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-socketio==5.11.2
apscheduler==3.10.4
pydantic-settings==2.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
mongomock-motor==0.0.21
redis==5.0.7
fakeredis==2.23.2
aiofiles==23.2.1
